+++ src/components/BoardBar/ExportWhiteboardDialogDownloadFile/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Nordeck IT + Consulting GmbH
+ * Copyright 2024 Nordeck IT + Consulting GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,4 @@
  * limitations under the License.
  */
 
-export { convertBlobToBase64 } from './convertBlobToBase64';
-export { determineImageSize } from './determineImageSize';
-export { findForegroundColor } from './findForegroundColor';
-export { isDefined } from './isDefined';
-export * from './matrix';
-export { FontsLoadedContextProvider, useFontsLoaded } from './useFontsLoaded';
-export { useLatestValue } from './useLatestValue';
-export { getUserColor } from './userColor';
+export { ExportWhiteboardDialogDownloadFile } from './ExportWhiteboardDialogDownloadFile';
